+++ src/main/java/com/yobombel/brewshare/stats/model/SpecificStats/GravityStats.java
@@ -1,8 +1,10 @@
 package com.yobombel.brewshare.stats.model.SpecificStats;
 
+import java.math.BigDecimal;
+
 public class GravityStats extends SpecificStats{
 
-    public GravityStats(double average, double max, double min) {
+    public GravityStats(BigDecimal average, BigDecimal max, BigDecimal min) {
         super(average, max, min);
     }
